Move ParamGen unit tests over to CIME's testing framework.
