--- CIME/ParamGen/tests/yaml_test_files/MOM_input.yaml
@@ -1,17 +1,0 @@
-Global:
-    INPUTDIR:
-        value: ${DIN_LOC_ROOT}/ocn/mom/${OCN_GRID}
-    RESTORE_SALINITY:
-        value:
-            $OCN_GRID == "tx0.66v1" and $COMP_ATM == "datm": True  # for C and G compsets on tx0.66v1
-            else: False
-    INIT_LAYERS_FROM_Z_FILE:
-        value:
-            $OCN_GRID == "gx1v6": True
-            $OCN_GRID == "tx0.66v1": True
-            $OCN_GRID == "tx0.25v1": True
-    TEMP_SALT_Z_INIT_FILE:
-        value:
-            $OCN_GRID == "gx1v6": "WOA05_pottemp_salt.nc"
-            $OCN_GRID == "tx0.66v1": "woa18_04_initial_conditions.nc"
-            $OCN_GRID == "tx0.25v1": "MOM6_IC_TS.nc"
